+++ src/common/TableContainer/Header/Header.module.css
@@ -1,0 +1,45 @@
+.header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  background-color: var(--white);
+  padding: 20px 20px 10px 20px;
+  z-index: 1;
+}
+
+.header_search {
+  position: relative;
+}
+
+.search_input {
+  width: 100%;
+  border: 1px solid var(--gray);
+  border-radius: 10px;
+  padding: 10px 20px;
+  font-weight: bold;
+  font-size: 15px;
+}
+
+.search_input_icon {
+  position: absolute;
+  top: 13px;
+  right: 20px;
+}
+
+
+/* Media queries */
+
+/* ========max-width: 800px ======== */
+@media only screen and (max-width: 800px) {
+  .search_input {
+    padding: 5px 20px;
+    font-size: 14px;
+  }
+
+  .search_input_icon {
+    top: 5px;
+    right: 15px;
+  }
+}
+
