--- src/common/TopTable/Table/Table.module.css
@@ -1,81 +1,0 @@
-.topTable {
-  width: 100%;
-}
-
-tbody tr:nth-child(2n) {
-  background-color: var(--lilac);
-}
-
-.rowProduct {
-  border-bottom: 1px solid transparent;
-  font-weight: bold;
-}
-
-.rowProduct:hover {
-  transition: 0.2s;
-  border-color: var(--gray);
-}
-
-.topTable td,
-.topTable th {
-  width: 20%;
-  text-align: center;
-  vertical-align: middle;
-  padding: 16px 0;
-}
-
-.itemProductBtn {
-  display: flex;
-  align-items: center;
-  margin: 0 auto;
-  padding: 8px 12px;
-  border-radius: 22px;
-}
-
-.itemProductBtn:hover {
-  background-color: var(--light-gray);
-  transition: 0.2s;
-}
-
-.wrapperImg,
-.imgProduct {
-  width: 32px;
-  height: 32px;
-}
-
-.wrapperImg {
-  margin-right: 8px;
-}
-
-.imgProduct {
-  object-fit: cover;
-  border-radius: 8px;
-}
-
-.indicatorYes {
-  background-color: var(--light-green);
-  color: var(--green);
-  display: inline-block;
-  border-radius: 22px;
-  padding: 8px 12px;
-  font-weight: bold;
-}
-
-.indicatorNo {
-  background-color: var(--light-orange);
-  color: var(--orange);
-  display: inline-block;
-  border-radius: 22px;
-  padding: 8px 12px;
-  font-weight: bold;
-}
-
-
-/* Media queries */
-
-/* ========max-width: 800px ======== */
-@media only screen and (max-width: 800px) {
-  tbody tr:nth-child(2n) {
-    background-color: var(--white);
-  }
-}
