+++ src/components/TableContainer/Buttons/ModalBtn/ExitButton.module.css
@@ -1,0 +1,5 @@
+.modal_product_control {
+  width: 100%;
+  text-align: right;
+  margin-bottom: 5px;
+}
