+++ src/components/TableContainer/Indicators/AvailableInStock/AvailableInStock.jsx
@@ -1,0 +1,10 @@
+import styles from "./AvailableInStock.module.css";
+
+
+const AvailableInStock = () => {
+  return (
+    <span className={styles.indicator_yes}>Есть на складе</span>
+  )
+};
+
+export default AvailableInStock;
