+++ src/components/TableContainer/Indicators/OutOfStock/OutOfStock.jsx
@@ -1,0 +1,10 @@
+import styles from "./OutOfStock.module.css";
+
+
+const OutOfStock = () => {
+  return (
+    <span className={styles.indicator_no}>Отсутствует</span>
+  )
+};
+
+export default OutOfStock;
