+++ src/components/TableContainer/Indicators/OutOfStock/OutOfStock.module.css
@@ -1,0 +1,8 @@
+.indicator_no {
+  background-color: var(--light-orange);
+  color: var(--orange);
+  display: inline-block;
+  border-radius: 22px;
+  padding: 8px 12px;
+  font-weight: bold;
+}
