--- src/components/TopTable/Buttons/ModalBtn/ExitButton.module.css
@@ -1,5 +1,0 @@
-.modalProductControl {
-  width: 100%;
-  text-align: right;
-  margin-bottom: 5px;
-}
