--- src/components/TopTable/Indicators/AvailableInStock/AvailableInStock.jsx
@@ -1,10 +1,0 @@
-import styles from "./AvailableInStock.module.css";
-
-
-const AvailableInStock = () => {
-  return (
-    <p className={styles.indicatorYes}>Есть на складе</p>
-  )
-};
-
-export default AvailableInStock;
