--- src/components/TopTable/Indicators/OutOfStock/OutOfStock.jsx
@@ -1,10 +1,0 @@
-import styles from "./OutOfStock.module.css";
-
-
-const OutOfStock = () => {
-  return (
-    <p className={styles.indicatorNo}>Отсутствует</p>
-  )
-};
-
-export default OutOfStock;
