--- src/components/TopTable/Indicators/OutOfStock/OutOfStock.module.css
@@ -1,8 +1,0 @@
-.indicatorNo {
-  background-color: var(--light-orange);
-  color: var(--orange);
-  display: inline-block;
-  border-radius: 22px;
-  padding: 8px 12px;
-  font-weight: bold;
-}
