+++ src/features/TableContainer/ModalWrapper/ModalProduct/ModalProduct.jsx
@@ -1,0 +1,15 @@
+import styles from "./ModalProduct.module.css";
+import ExitButton from "../../../../components/TableContainer/Buttons/ModalBtn/ExitButton";
+import ModalProductContent from "./ModalProductContent/ModalProductContent";
+
+
+const ModalProduct = ({hideModal, dataForModal}) => {
+  return (
+    <div className={styles.modal_product_card}>
+      <ExitButton hideModal={hideModal} />
+      <ModalProductContent dataForModal={dataForModal} />
+    </div>
+  )
+};
+
+export default ModalProduct;
