+++ src/features/TableContainer/ModalWrapper/ModalProduct/ModalProductContent/ModalProductContent.module.css
@@ -1,0 +1,32 @@
+.modal_product_title,
+.modal_product_price,
+.modal_product_count {
+  margin-bottom: 15px;
+}
+
+.modal_wrapper_img {
+  max-width: 310px;
+  height: 219px;
+  margin-bottom: 15px;
+}
+
+.modal_img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.modal_product_price,
+.modal_product_count,
+.modal_product_link {
+  font-weight: bold;
+}
+
+.modal_product_description {
+  margin-bottom: 12px;
+  line-height: 17px;
+}
+
+.modal_product_link {
+  text-decoration-line: underline;
+}
