+++ src/features/TableContainer/ModalWrapper/ModalWrapper.module.css
@@ -1,0 +1,16 @@
+.modal_wrapper {
+  display: flex;
+  width: 100%;
+  position: fixed;
+  top: 0;
+  background-color: var(--foggy);
+  height: 100vh;
+  z-index: 2;
+  align-items: center;
+  justify-content: center;
+  padding: 0 30px;
+}
+
+.modal_wrapper_hide {
+  display: none;
+}
