+++ src/features/TableContainer/Table/TableBody/TableBody.module.css
@@ -1,0 +1,8 @@
+/* Media queries */
+
+/* ========max-width: 800px ======== */
+@media only screen and (max-width: 800px) {
+  .table_body {
+    display: none;
+  }
+}
