--- src/features/TopTable/ModalWrapper/ModalProduct/ModalProduct.module.css
@@ -1,7 +1,0 @@
-.modalProductCard {
-  max-width: 350px;
-  padding: 20px;
-  background: var(--white);
-  border-radius: 10px;
-  box-shadow: 0px 0px 22px 12px rgba(34, 60, 80, 0.2);
-}
