--- src/features/TopTable/ModalWrapper/ModalProduct/ModalProductContent/ModalProductContent.module.css
@@ -1,35 +1,0 @@
-.modalProductTitle {
-  margin-bottom: 15px;
-}
-
-.modalImgWrapper {
-  max-width: 310px;
-  height: 219px;
-  margin-bottom: 15px;
-}
-
-.modalImg {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.modalProductPrice {
-  font-weight: bold;
-  margin-bottom: 15px;
-}
-
-.modalProductDescription {
-  margin-bottom: 12px;
-  line-height: 17px;
-}
-
-.modalProductCount {
-  font-weight: bold;
-  margin-bottom: 15px;
-}
-
-.modalProductLink {
-  font-weight: bold;
-  text-decoration-line: underline;
-}
