--- src/features/TopTable/ModalWrapper/ModalWrapper.module.css
@@ -1,16 +1,0 @@
-.modalWrapper {
-  display: flex;
-  width: 100%;
-  position: fixed;
-  top: 0;
-  background-color: var(--foggy);
-  height: 100vh;
-  z-index: 2000;
-  align-items: center;
-  justify-content: center;
-  padding: 0 30px;
-}
-
-.hideModal {
-  display: none;
-}
