--- src/features/TopTable/Table/Tbody/Product/Product.module.css
@@ -1,9 +1,0 @@
-td {
-  width: 20%;
-  vertical-align: middle;
-  padding: 16px 0;
-}
-
-.rowProduct {
-  font-weight: bold;
-}
