--- src/features/TopTable/Table/Tbody/Tbody.module.css
@@ -1,8 +1,0 @@
-/* Media queries */
-
-/* ========max-width: 800px ======== */
-@media only screen and (max-width: 800px) {
-  .tbody {
-    display: none;
-  }
-}
