--- src/features/TopTable/Table/Thead/Thead.module.css
@@ -1,33 +1,0 @@
-th {
-  width: 20%;
-  text-align: center;
-  vertical-align: middle;
-  padding: 16px 0;
-}
-
-.topTableTitles {
-  position: fixed;
-  top: 70px;
-  left: 0;
-  right: 0;
-  display: flex;
-  background-color: var(--white);
-  font-size: 18px;
-  font-weight: bold;
-  z-index: 1000;
-}
-
-.dateFilterBtn,
-.priceFilterBtn {
-  margin-right: 8px;
-}
-
-
-/* Media queries */
-
-/* ========max-width: 800px ======== */
-@media only screen and (max-width: 800px) {
-  .thead {
-    display: none;
-  }
-}
